fix type warnings

+++ libheif/api/libheif/heif.cc
@@ -1584,7 +1584,7 @@ const uint8_t* heif_image_get_plane_readonly(const struct heif_image* image,
   uint32_t stride;
   const auto* p = image->image->get_plane(channel, &stride);
 
-  if (stride > std::numeric_limits<int>::max()) {
+  if (stride > static_cast<uint32_t>(std::numeric_limits<int>::max())) {
     return nullptr;
   }
 
@@ -1609,7 +1609,7 @@ uint8_t* heif_image_get_plane(struct heif_image* image,
   uint32_t stride;
   uint8_t* p = image->image->get_plane(channel, &stride);
 
-  if (stride > std::numeric_limits<int>::max()) {
+  if (stride > static_cast<uint32_t>(std::numeric_limits<int>::max())) {
     return nullptr;
   }
